[Foundation] Move the WrappedNSInputStream out of the handler file.
