+++ src/Foundation/NSUrlSessionHandler.cs
@@ -27,7 +27,6 @@
 
 using System;
 using System.Collections.Generic;
-using System.Globalization;
 using System.IO;
 using System.Net;
 using System.Net.Http;
@@ -1109,101 +1108,6 @@ namespace Foundation {
 			}
 		}
 
-		//
-		// Copied from https://github.com/mono/mono/blob/2019-02/mcs/class/System.Net.Http/System.Net.Http/StreamContent.cs.
-		//
-		// This is not a perfect solution, but the most robust and risk-free approach.
-		//
-		// The implementation depends on Mono-specific behavior, which makes SerializeToStreamAsync() cancellable.
-		// Unfortunately, the CoreFX implementation of HttpClient does not support this.
-		//
-		// By copying Mono's old implementation here, we ensure that we're compatible with both HttpClient implementations,
-		// so when we eventually adopt the CoreFX version in all of Mono's profiles, we don't regress here.
-		//
-		class MonoStreamContent : HttpContent
-		{
-			readonly Stream content;
-			readonly int bufferSize;
-			readonly CancellationToken cancellationToken;
-			readonly long startPosition;
-			bool contentCopied;
-
-			public MonoStreamContent (Stream content)
-				: this (content, 16 * 1024)
-			{
-			}
-
-			public MonoStreamContent (Stream content, int bufferSize)
-			{
-				if (content is null)
-					ObjCRuntime.ThrowHelper.ThrowArgumentNullException (nameof (content));
-
-				if (bufferSize <= 0)
-					ObjCRuntime.ThrowHelper.ThrowArgumentOutOfRangeException (nameof (bufferSize), bufferSize, "Buffer size must be >0");
-
-				this.content = content;
-				this.bufferSize = bufferSize;
-
-				if (content.CanSeek) {
-					startPosition = content.Position;
-				}
-			}
-
-			//
-			// Workarounds for poor .NET API
-			// Instead of having SerializeToStreamAsync with CancellationToken as public API. Only LoadIntoBufferAsync
-			// called internally from the send worker can be cancelled and user cannot see/do it
-			//
-			internal MonoStreamContent (Stream content, CancellationToken cancellationToken)
-				: this (content)
-			{
-				// We don't own the token so don't worry about disposing it
-				this.cancellationToken = cancellationToken;
-			}
-
-			protected override Task<Stream> CreateContentReadStreamAsync ()
-			{
-				return Task.FromResult (content);
-			}
-
-			protected override void Dispose (bool disposing)
-			{
-				if (disposing) {
-					content.Dispose ();
-				}
-
-				base.Dispose (disposing);
-			}
-
-			protected override Task SerializeToStreamAsync (Stream stream, TransportContext? context)
-			{
-				if (contentCopied) {
-					if (!content.CanSeek) {
-						throw new InvalidOperationException ("The stream was already consumed. It cannot be read again.");
-					}
-
-					content.Seek (startPosition, SeekOrigin.Begin);
-				} else {
-					contentCopied = true;
-				}
-
-				return content.CopyToAsync (stream, bufferSize, cancellationToken);
-			}
-
-#if !NET
-			internal
-#endif
-			protected override bool TryComputeLength (out long length)
-			{
-				if (!content.CanSeek) {
-					length = 0;
-					return false;
-				}
-				length = content.Length - startPosition;
-				return true;
-			}
-		}
-
 		class NSUrlSessionDataTaskStream : Stream
 		{
 			readonly Queue<NSData> data;
@@ -1348,120 +1252,6 @@ namespace Foundation {
 				throw new InvalidOperationException ();
 			}
 		}
-
-		class WrappedNSInputStream : NSInputStream
-		{
-			NSStreamStatus status;
-			CFRunLoopSource source;
-			readonly Stream stream;
-			bool notifying;
-
-			public WrappedNSInputStream (Stream inputStream)
-			{
-				status = NSStreamStatus.NotOpen;
-				stream = inputStream;
-				source = new CFRunLoopSource (Handle, false);
-			}
-
-			public override NSStreamStatus Status => status;
-
-			public override void Open ()
-			{
-				status = NSStreamStatus.Open;
-				Notify (CFStreamEventType.OpenCompleted);
-			}
-
-			public override void Close ()
-			{
-				status = NSStreamStatus.Closed;
-			}
-
-			public override nint Read (IntPtr buffer, nuint len)
-			{
-				var sourceBytes = new byte [len];
-				var read = stream.Read (sourceBytes, 0, (int)len);
-				Marshal.Copy (sourceBytes, 0, buffer, (int)len);
-
-				if (notifying)
-					return read;
-
-				notifying = true;
-				if (stream.CanSeek && stream.Position == stream.Length) {
-					Notify (CFStreamEventType.EndEncountered);
-					status = NSStreamStatus.AtEnd;
-				}
-				notifying = false;
-
-				return read;
-			}
-
-			public override bool HasBytesAvailable ()
-			{
-				return true;
-			}
-
-			protected override bool GetBuffer (out IntPtr buffer, out nuint len)
-			{
-				// Just call the base implemention (which will return false)
-				return base.GetBuffer (out buffer, out len);
-			}
-
-			// NSInvalidArgumentException Reason: *** -propertyForKey: only defined for abstract class.  Define -[System_Net_Http_NSUrlSessionHandler_WrappedNSInputStream propertyForKey:]!
-			protected override NSObject? GetProperty (NSString key)
-			{
-				return null;
-			}
-
-			protected override bool SetProperty (NSObject? property, NSString key)
-			{
-				return false;
-			}
-
-			protected override bool SetCFClientFlags (CFStreamEventType inFlags, IntPtr inCallback, IntPtr inContextPtr)
-			{
-				// Just call the base implementation, which knows how to handle everything.
-				return base.SetCFClientFlags (inFlags, inCallback, inContextPtr);
-			}
-
-#if NET
-			public override void Schedule (NSRunLoop aRunLoop, NSString nsMode)
-#else
-			public override void Schedule (NSRunLoop aRunLoop, string mode)
-#endif
-			{
-				var cfRunLoop = aRunLoop.GetCFRunLoop ();
-#if !NET
-				var nsMode = new NSString (mode);
-#endif
-
-				cfRunLoop.AddSource (source, nsMode);
-
-				if (notifying)
-					return;
-
-				notifying = true;
-				Notify (CFStreamEventType.HasBytesAvailable);
-				notifying = false;
-			}
-
-#if NET
-			public override void Unschedule (NSRunLoop aRunLoop, NSString nsMode)
-#else
-			public override void Unschedule (NSRunLoop aRunLoop, string mode)
-#endif
-			{
-				var cfRunLoop = aRunLoop.GetCFRunLoop ();
-#if !NET
-				var nsMode = new NSString (mode);
-#endif
-
-				cfRunLoop.RemoveSource (source, nsMode);
-			}
-
-			protected override void Dispose (bool disposing)
-			{
-				stream?.Dispose ();
-			}
-		}
 	}
+
 }
